First attempt at adding support for Windows Nodes in the broker.
Two gems have been modified: openshift-origin-controller and openshift-origin-msg-broker-mcollective

The application container proxy now exposes a 'kernel' filtering parameter. Filtering is done on the 'kernel' mcollective fact, and it should be either Linux or Windows.

This is used up the stack to list cartridges from both Linux and Windows, and to properly pick the right nodes when creating apps. If an app's cartridge has a 'windows' category assigned, the gear created for that will be placed on a Windows machine.

Change-Id: Id28ff34d07da3910825dcb39644c01eac82919bb